Fixed bugs in span deserialization.

+++ Vostok.Hercules.Client.TracingApi/Dto/SpanDto.cs
@@ -18,13 +18,13 @@ namespace Vostok.Hercules.Client.TracingApi.Dto
         public long BeginTimestampUtc;
 
         [JsonProperty("beginTimestampUtcOffset")]
-        public int BeginTimestampUtcOffset;
+        public long BeginTimestampUtcOffset;
 
         [JsonProperty("endTimestampUtc")]
         public long EndTimestampUtc;
 
         [JsonProperty("endTimestampUtcOffset")]
-        public int EndTimestampUtcOffset;
+        public long EndTimestampUtcOffset;
 
         [JsonProperty("annotations")]
         public Dictionary<string, string> Annotations;
